requirements.txt
pyTelegramBotAPI==4.12.0
Flask==3.0.0
requests==2.32.4
python-dotenv==1.0.1
Flask-Cors==4.0.1